+++ public/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>6 cities</title>
   <link rel="stylesheet" href="/css/main.css">
-  <link rel="stylesheet" href="/css/additions.css"
+  <link rel="stylesheet" href="/css/additions.css">
   <link rel="stylesheet" href="/css/loader.css">
   <link rel="stylesheet" href="https://unpkg.com/leaflet@1.5.1/dist/leaflet.css"
   integrity="sha512-xwE/Az9zrjBIphAcBb3F6JVqxf46+CDLwfLMHloNu6KEQCAWi6HcDUbeOfBIptF7tcCzusKFjFw2yuvEpDL9wQ=="
