+++ src/components/review-form/review-form.test.js
@@ -2,21 +2,25 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 
 import {ReviewForm} from './review-form.jsx';
-
-const errorStyle = {color: `#BF616A`, paddingBottom: `5px`};
+const TEXT = `Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
+tempor incididunt ut labore et dolore magna aliqua.`;
 
 it(`ReviewForm renders correctly`, () => {
 
   const tree = renderer
     .create(
         <ReviewForm
-          onSubmit={() => {}}
+          offerId={1}
+          sending={false}
+          error={false}
+          rating={4}
+          isRatingValid={true}
+          text={TEXT}
+          isTextValid={true}
           onRatingChange={() => {}}
           onTextChange={() => {}}
-          errorStyle={errorStyle}
-          textRef={React.createRef()}
-          submitRef={React.createRef()}
-          formRef={React.createRef()}
+          onSetSendingFlag={() => {}}
+          onSubmitReview={() => {}}
         />, {
           createNodeMock: () => document.createElement(`div`)
         })
